strace: Fix strace pre-build step
Summary:
Just like musl, the `POST_PATCH_HOOK` is not executed when using a
`local.mk`. So, we move the operation to the `PRE_CONFIGURE_HOOK` of
strace. With this we can get rid of the annoying step in our `build.rb`
`pre_build`, which is run everytime we compile buildroot, even if strace
is already compiled and doesn't need a re-compilation.

Ref T19412

Test Plan: CI

Reviewers: jmaselbas, ysionneau, O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T19412

Differential Revision: https://phab.kalray.eu/D24479