gcc: We need to set BR2_TOOLCHAIN_GCC_AT_LEAST_12 in order for GDB to compile
Summary:
When using a custom version the flag `BR2_TOOLCHAIN_GCC_AT_LEAST_XX` is
not set, because we can't know which version this "custom version" is.
So, GDB can't be compiled for the target. So, we just set the minimum
necessary version, because we can be sure the "custom version" is at
least a GCC 4.8, which is fortunately enough for GDB.

Ref T22509

Test Plan: CI

Reviewers: O53 Buildroot Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T22509

Differential Revision: https://phab.kalray.eu/D34833